+++ packages/native/src/react-native-video.ts
@@ -1,2 +1,5 @@
-export { default } from 'react-native-video';
+import Video from 'react-native-video';
+
 export * from 'react-native-video';
+export { Video };
+export default Video;