+++ packages/react-native/src/rn-polyfills/index.ts
@@ -1,7 +1,0 @@
-import { setup as setupSymbolAsyncIterator } from './symbol-asynciterator';
-import { setup as setupURLPolyfill } from './url';
-
-export function setup() {
-  setupSymbolAsyncIterator();
-  setupURLPolyfill();
-}