+++ packages/react-native/src/router/types/RouteScreen.ts
@@ -1,3 +1,4 @@
+import { NativeStackNavigationOptions } from '@granite-js/native/@react-navigation/native-stack';
 import { Screen } from './Screen';
 
 /**
@@ -14,4 +15,9 @@ export interface RouteScreen {
    * @description Screen component
    */
   component: Screen;
+  /**
+   * @name screenOptions
+   * @description Screen options for React Navigation (can be static or a function that receives route params)
+   */
+  screenOptions?: NativeStackNavigationOptions;
 }